+++ NEWS.md
@@ -4,6 +4,7 @@
 
 * `colFormat(percent = TRUE)` now works correctly when viewing tables in IE11.
 * Table searching now works when row selection is enabled.
+* In `colDef()`, R functions for `class` and `style` now always receive cell values as single elements.
 
 # reactable 0.1.0